Move experiences from modal to main page
