+++ src/components/Modal.jsx
@@ -3,7 +3,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {faClose, faPersonBiking, faPersonRunning, faPersonWalking, faChild, faWeightHanging} from '@fortawesome/free-solid-svg-icons';
 import PropTypes from 'prop-types';
 
-const Modal = ({skill, data, handleModal}) => {
+const Modal = ({skill, handleModal}) => {
   const skills = {
     "master": "Master",
     "expert": "Expert",
@@ -20,10 +20,6 @@ const Modal = ({skill, data, handleModal}) => {
     "no-experience-interested": faChild
   };
 
-  const jobExperience = data.experiences.filter((job) => {
-    return job.category === 'jobs';
-  });
-
   return (
     <section className="modal flexColumn">
       <nav className="navModal flex">
@@ -35,25 +31,12 @@ const Modal = ({skill, data, handleModal}) => {
         <p className="weightLight"><span className="textGray">Recomendations: </span>{skill.recommendations}</p>
         <p className="weightLight"><span className="textGray">Weight: </span><FontAwesomeIcon icon={faWeightHanging} /> {skill.weight?.toFixed(1)}</p>
       </div>
-      <div className="modalInfo flexColumn">
-        <h4 className="weightLight">{`${data.person?.name}'s experiences:`}</h4>
-        {jobExperience.map((job) => {
-          return (
-            <div key={job.id}>
-              <p className="weightNormal textYellow">{job.name}</p>
-              <p className="weightNormal textGray">{job.organizations[0].name}</p>
-              <p className="weightNormal textGray">{job.fromMonth} {job.fromYear} - {job.toMonth ? job.toMonth : "Now"} {job.toYear}</p>
-            </div>
-          );
-        })}
-      </div>
     </section>
   );
 };
 
 Modal.propTypes = {
   skill: PropTypes.node.isRequired,
-  data: PropTypes.node.isRequired,
   handleModal: PropTypes.node.isRequired,
 }
