+++ src/stylesheets/components/Experiences.css
@@ -1,0 +1,4 @@
+.experiences {
+  gap: 1rem;
+  width: 90%;
+}
